Split up the main entry point between electron and non-electron.
