+++ src-electron/main-process/main.ts
@@ -15,14 +15,11 @@
  *    limitations under the License.
  */
 
-const { app } = require('electron')
-
 // enable stack trace to be mapped back to the correct line number in TypeScript source files.
 require('source-map-support').install()
 
 import * as args from '../util/args'
 const env = require('../util/env')
-const windowJs = require('../ui/window')
 const startup = require('./startup')
 
 env.versionsCheck()
@@ -33,66 +30,9 @@ if (process.env.DEV) {
   env.setProductionEnv()
 }
 
-function hookSecondInstanceEvents(argv: args.Arguments) {
-  app.whenReady().then(() => startup.startUpSecondaryInstance(app.quit, argv))
-}
-
-/**
- * Hook up all the events for the electron app object.
- */
-function hookMainInstanceEvents(argv: args.Arguments) {
-  app.whenReady().then(() => startup.startUpMainInstance(app.quit, argv))
-
-  if (!argv._.includes('server')) {
-    app.on('window-all-closed', () => {
-      if (process.platform !== 'darwin') {
-        app.quit()
-      }
-    })
-    app.on('activate', () => {
-      env.logInfo('Activate...')
-      windowJs.windowCreateIfNotThere(argv.httpPort)
-    })
-  }
-
-  app.on('will-quit', () => {
-    startup.shutdown()
-  })
-
-  app.on(
-    'second-instance',
-    (event: Event, commandLine: string[], workingDirectory: string) => {
-      env.logInfo(`Zap instance started with command line: ${commandLine}`)
-    }
-  )
-}
-
-// Main lifecycle of the application
-if (app != null) {
-  let argv = args.processCommandLineArguments(process.argv)
-  let reuseZapInstance = argv.reuseZapInstance
-  let canProceedWithThisInstance
-  let gotLock = app.requestSingleInstanceLock()
-
-  if (reuseZapInstance) {
-    canProceedWithThisInstance = gotLock
-  } else {
-    canProceedWithThisInstance = true
-  }
-  if (canProceedWithThisInstance) {
-    hookMainInstanceEvents(argv)
-  } else {
-    // The 'second-instance' event on app was triggered, we need
-    // to quit.
-    hookSecondInstanceEvents(argv)
-  }
-} else {
-  // If the code is executed via 'node' and not via 'electron', then this
-  // is where we end up.
-  startup.startUpMainInstance(
-    null,
-    args.processCommandLineArguments(process.argv)
-  )
-}
-
-exports.loaded = true
+// If the code is executed via 'node' and not via 'electron', then this
+// is where we end up.
+startup.startUpMainInstance(
+  null,
+  args.processCommandLineArguments(process.argv)
+)
